make isolated compoent out of Stepper
